Rearrange tests so they make sense
